[rename] coupons : coupon, usercoupon repository 위치 변경 #15
