+++ coupon/src/main/java/com/haot/coupon/infrastructure/repository/CouponRepository.java
@@ -1,4 +1,4 @@
-package com.haot.coupon.infrastructure;
+package com.haot.coupon.infrastructure.repository;
 
 import com.haot.coupon.domain.model.Coupon;
 import org.springframework.data.jpa.repository.JpaRepository;
